refactor: Move all source code into 'src/'
The index.ts now lives among the rest of the source code in 'src/',
instead of weirdly being placed among tool configs.
