--- index.ts
@@ -1,2 +1,0 @@
-export * from './lib/express-promises.js'
-export * from './lib/termination.js'
