+++ package.json
@@ -11,10 +11,10 @@
   "scripts": {
     "clean": "rimraf dist",
     "build": "npm run clean && tsc",
-    "lint": "eslint --ignore-path .gitignore . && tsc --noEmit",
-    "lint-fix": "eslint --fix --ignore-path .gitignore . && tsc --noEmit",
+    "lint": "tsc --noEmit && eslint --ignore-path .gitignore .",
+    "lint-fix": "tsc --noEmit && eslint --fix --ignore-path .gitignore .",
     "test": "mocha --recursive \"test/**/*.test.*\"",
-    "coverage": "c8 --extension=.ts --all --reporter=text --reporter=lcov npm test"
+    "coverage": "c8 --all --src=src --reporter=text --reporter=lcov npm test"
   },
   "repository": {
     "type": "git",
