+++ tsconfig.eslint.json
@@ -1,7 +1,7 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": [
-    "dist",
-    "node_modules"
+  "include": [
+    "src",
+    "test"
   ]
 }
