+++ tsconfig.json
@@ -9,9 +9,7 @@
     "declaration": true,
     "outDir": "./dist"
   },
-  "exclude": [
-    "test",
-    "dist",
-    "node_modules"
+  "include": [
+    "src"
   ]
 }
